+++ crates/pathfinder/src/consensus/inner/p2p_task/p2p_task_tests.rs
@@ -105,6 +105,7 @@ impl TestEnvironment {
             finalized_blocks,
             // Only used for failure injection, which does not happen in these tests
             &PathBuf::default(),
+            pathfinder_compiler::ResourceLimits::recommended(),
             true,
             None,
             None,